Refactor UI: Centralize rendering logic to MapRenderUtils and NetworkRenderer
